Chargeのテストケースの完成
